homework/stage05_data-storage/requirements.txt
accelerate==0.18.0
anyio==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.3
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.16.1
executing==2.2.0
fastapi==0.95.1
fastjsonschema==2.21.2
filelock==3.12.0
fonttools==4.59.1
fqdn==1.5.1
frozendict==2.4.6
fsspec==2023.4.0
gcloud==0.17.0
googleapis-common-protos==1.59.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.5.0
httpx==0.28.1
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
lxml==6.0.0
MarkupSafe==2.1.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mistune==3.1.3
multitasking==0.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.2
oauth2client==3.0.0
overrides==7.7.0
packaging==23.1
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
peewee==3.18.2
Pillow==9.5.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==4.23.0rc2
psutil==5.9.5
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==1.10.7
Pygments==2.19.2
pyparsing==3.0.9
Pyrebase==3.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==3.3.0
python-jwt==2.0.1
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0
pyzmq==27.0.1
referencing==0.36.2
regex==2023.3.23
requests==2.32.5
requests-toolbelt==0.7.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.0
rsa==4.9
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.7
stack-data==0.6.3
starlette==0.26.1
terminado==0.18.1
tinycss2==1.4.0
torch==2.0.0
tornado==6.5.2
tqdm==4.65.0
traitlets==5.14.3
transformers==4.28.1
types-python-dateutil==2.9.0.20250809
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.22.0
watchfiles==0.19.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
yfinance==0.2.65
z3==0.2.0
